requirements.txt
fastapi
numpy
pandas
matplotlib
requests
beautifulsoup4
pillow
google-genai
dotenv
subprocess.run
fastparquet